fix the hospitalization prediction in Texas
The projection of the incident hospitalization is generated largely based on the data in past one month (Jun 4~ Jul 4). However, the confirmed cases in TX increases rapidly during the last month, making the training process unstable and thus lead to over-estimated new hospitalization cases (as well as the recover rate of current hospitalization cases). We have fixed this issue by adding constraints on the model parameters to stablize the training and prevent from over-estimation. We have updated the hospitalization projections in TX, which we believe is in the reasonable range (1000-2000).