Added code to pull data into this folder
cases.csv is a file of daily inciddent cases
deaths.csv is a file of daily deaths
hosps.csv is a file of daily hosps

An important file is threestreams.csv. This file contains columns that records daily  cases, deaths, and hosps all in a single csv.
The next commit fill contain a script where you can begin the runningmean work.

-prof m