Re-ordering push to github and push to zoltar
Having push to zoltar happening last so that if data are updated for targets that are not part of zoltar, the zoltar upload (which would be entirely duplicate data) doesn't fail and cancel the github push. Now the push of any data would happen to GitHub successfully but it still could fail.